requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
openai==1.3.5
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
httpx==0.25.2
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1
requests==2.31.0
aiofiles==23.2.1
loguru==0.7.2
# Vector Database Dependencies
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
beautifulsoup4==4.12.2
scrapy==2.11.0
# Security & Compliance Dependencies
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.2.1 